% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{accuracyComplexityPlot}
\alias{accuracyComplexityPlot}
\title{Title}
\usage{
accuracyComplexityPlot(ReportTable, AdjFactor = 1, Main = "",
  Mode = "LocMin", Xlims = NULL)
}
\arguments{
\item{Xlims}{}
}
\description{
Title
}
