% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_operations.R
\name{AddNode2Node}
\alias{AddNode2Node}
\title{Adds a node to each graph node}
\usage{
AddNode2Node(X, NodePositions, ElasticMatrix, Partition)
}
\arguments{
\item{ElasticMatrix}{}
}
\description{
This grammar operation adds a node to each graph node. The positions of the node
is chosen as a linear extrapolation for a leaf node (in this case the elasticity of
a newborn star is chosed as in BisectEdge operation), or as the data point giving
the minimum local MSE for a star (without any optimization).
}
\details{

}
