% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMAdjustment.R
\name{AdjustByConstant}
\alias{AdjustByConstant}
\title{Title}
\usage{
AdjustByConstant(GraphInfo, k = 2, divMu = 10, divLambda = 10)
}
\arguments{
\item{GraphInfo}{}

\item{k}{}

\item{divMu}{}

\item{divLambda}{}
}
\description{
Title
}
