% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_operations.R
\name{ApplyOptimalGraphGrammarOpeation}
\alias{ApplyOptimalGraphGrammarOpeation}
\title{Application of the grammar operation. This in an internal function that should not be used in by the end-user}
\usage{
ApplyOptimalGraphGrammarOpeation(X, NodePositions, ElasticMatrix,
  AdjustVect = NULL, operationtypes, SquaredX = NULL, verbose = FALSE,
  n.cores = 1, EnvCl = NULL, MinParOP = 20, MaxNumberOfIterations = 100,
  eps = 0.01, TrimmingRadius = Inf, Mode = 1, FinalEnergy = "Base",
  alpha = 1, beta = 1, gamma = 1, FastSolve = FALSE, EmbPointProb = 1,
  AvoidSolitary = FALSE, AdjustElasticMatrix = NULL, ...)
}
\arguments{
\item{X}{numerical 2D matrix, the n-by-m matrix with the position of n m-dimensional points}

\item{NodePositions}{numerical 2D matrix, the k-by-m matrix with the position of k m-dimensional points}

\item{ElasticMatrix}{numerical 2D matrix, the k-by-k elastic matrix}

\item{AdjustVect}{}

\item{operationtypes}{string vector containing the operation to use}

\item{SquaredX}{rowSums(X^2), if NULL it will be computed}

\item{verbose}{boolean. Should addition information be displayed}

\item{n.cores}{integer. How many cores to use. If EnvCl is not NULL, that cliuster setup will be used,
otherwise a SOCK cluster willbe used}

\item{EnvCl}{a cluster structure returned, e.g., by makeCluster.
If a cluster structure is used, all the nodes must be able to access all the variable needed by PrimitiveElasticGraphEmbedment}

\item{MinParOP}{integer, the minimum number of operations to use parallel computation}

\item{MaxNumberOfIterations}{is an integer number indicating the maximum number of iterations for the EM algorithm}

\item{eps}{a real number indicating the minimal relative change in the nodenpositions
to be considered the graph embedded (convergence criteria)}

\item{TrimmingRadius}{is a real value indicating the trimming radius, a parameter required for robust principal graphs
(see https://github.com/auranic/Elastic-principal-graphs/wiki/Robust-principal-graphs)}

\item{Mode}{integer, the energy mode. It can be 1 (difference is computed using the position of the nodes) and
2 (difference is computed using the changes in elestic energy of the configuraztions)}

\item{FinalEnergy}{string indicating the final elastic emergy associated with the configuration. Currently it can be "Base" or "Penalized"}

\item{alpha}{positive numeric, the value of the alpha parameter of the penalized elastic energy}

\item{beta}{positive numeric, the value of the beta parameter of the penalized elastic energy}

\item{gamma}{}

\item{FastSolve}{boolean, should FastSolve be used when fitting the points to the data?}

\item{EmbPointProb}{numeric between 0 and 1. If less than 1 point will be sampled at each iteration. Prob indicate the probability of
using each points. This is an *experimental* feature, which may helps speeding up the computation if a large number of points is present.}

\item{AvoidSolitary}{boolean, should configurations with "solitary nodes", i.e., nodes without associted points be discarded?}

\item{AdjustElasticMatrix}{}

\item{...}{}
}
\value{

}
\description{
Application of the grammar operation. This in an internal function that should not be used in by the end-user
}
