% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.structure.R
\name{BranchingDimension}
\alias{BranchingDimension}
\title{Produce a multidimensional dimension matrix}
\usage{
BranchingDimension(X, PrintGraph, Start = NULL)
}
\arguments{
\item{X}{a data matrix}

\item{PrintGraph}{an ElPiGraph object}

\item{Start}{the Starting node. If NULL (the default), a random end point will be selected}
}
\value{
a list with two elements
\itemize{
 \item Branches, a list containing the (ordered) nodes composing the different branches
 \item DimMatrix, a matrix contianing the normalized position of points acoross branches. The columns
 correpsond to the different branches definezd by Branches. If a points is not assigned to a branch NA is reported.
}
}
\description{
Produce a multidimensional dimension matrix
}
