% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{circle_data}
\alias{circle_data}
\title{Points on a circle}
\format{A data frame with 200 rows and 3 variables}
\description{
An example dataset containing 3d points distribued on a circle
}
\keyword{data}
