% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlterStructure.R
\name{CollapseBrances}
\alias{CollapseBrances}
\title{Filter "small" branches}
\usage{
CollapseBrances(X, TargetPG, Mode = "PointNumber", ControlPar = 5,
  TrimmingRadius = Inf, PlotSelected = TRUE)
}
\arguments{
\item{X}{numeric matrix, the data matrix}

\item{TargetPG}{list, the ElPiGraph structure to extend}

\item{Mode}{string, the mode used to extend the graph. "PointNumber", "PointNumber_Extrema", "PointNumber_Leaves",
"EdgesNumber", and "EdgesLength" are currently implemented}

\item{ControlPar}{positive numeric, the paramter used to control the contribution of the different data points}

\item{TrimmingRadius}{positive numeric, the trimming radius used to control distance}

\item{PlotSelected}{boolean, should a diagnostic plot be visualized (currently not implemented)}
}
\value{
a list with 2 values: Nodes (a matrix containing the new nodes positions) and Edges (a matrix describing the new edge structure)

The value of ControlPar has a different interpretation depending on the valus of Mode.

If Mode = "PointNumber", branches with less that ControlPar points projected on the branch
(points projected on the extreme points are not considered) are removed

If Mode = "PointNumber_Extrema", branches with less that ControlPar points projected on the branch or the extreme
points are removed

If Mode = "PointNumber_Leaves", branches with less that ControlPar points projected on the branch and any leaf points
(points projected on non-leaf extreme points are not considered) are removed

If Mode = "EdgesNumber", branches with less that ControlPar edges are removed

If Mode = "EdgesLength", branches with with a length smaller than ControlPar are removed
}
\description{
Filter "small" branches
}
