% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlterStructure.R
\name{CollapseCliques}
\alias{CollapseCliques}
\title{Filter cliques}
\usage{
CollapseCliques(TargetPG, MaxClZize = NULL, DistThr = NULL,
  Compensate = 3)
}
\arguments{
\item{TargetPG}{}

\item{MaxClZize}{}

\item{DistThr}{}
}
\description{
Filter cliques
}
