% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{CompareExpOnBranches}
\alias{CompareExpOnBranches}
\title{Title}
\usage{
CompareExpOnBranches(ExpData, Paths, TargetPG, Partition, PrjStr, Main = "",
  genes = 4)
}
\arguments{
\item{ExpData}{}

\item{Paths}{}

\item{TargetPG}{}

\item{Partition}{}

\item{PrjStr}{}

\item{Main}{}

\item{genes}{}
}
\description{
Title
}
