% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{CompareOnBranches}
\alias{CompareOnBranches}
\title{Compare feature values over different branches}
\usage{
CompareOnBranches(X, Paths, TargetPG, Partition, PrjStr, Main = "",
  Features = 4)
}
\arguments{
\item{X}{the data matrix, note that features must be names, i.e. column names must be non-empty}

\item{Paths}{a list of paths used to construct the pseudotime. Note that the beginning of the pseudotime
will be placed in correspondence of the fisrst node for each path. Hence the comparison is meaningfull
only if all the path begin from the same node (Although, this will not be enforced by the function)}

\item{TargetPG}{the ElPiGraph structure to be used}

\item{Partition}{A partition vactor attributing each point to at most one node}

\item{PrjStr}{A projecturion structure, as produced by project_point_onto_graph}

\item{Main}{string, the main title of the plot}

\item{Features}{either a string vector or a positive integer. If a string vector, it will be interpreted as
the name of the features to plot (matched to the colnames of X). If positive integer, it will be interpreted
as the number of the topmost vayring features of plot}
}
\description{
Compare feature values over different branches
}
