% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ComputeSpringLaplacianMatrix}
\alias{ComputeSpringLaplacianMatrix}
\title{Compute the Laplacian matrix}
\usage{
ComputeSpringLaplacianMatrix(ElasticMatrix)
}
\arguments{
\item{ElasticMatrix}{an e-by-e elastic matrix}
}
\value{
the Laplacian matrix
}
\description{
Compute the Laplacian matrix
}
