% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ConstructGraph}
\alias{ConstructGraph}
\title{Generate an igraph object from a ElPiGraph structure}
\usage{
ConstructGraph(PrintGraph)
}
\arguments{
\item{PrintGraph}{A principal graph object}
}
\value{
An igraph network
}
\description{
Generate an igraph object from a ElPiGraph structure
}
