% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVSEdges.R
\name{CrossEmbedment}
\alias{CrossEmbedment}
\title{Cross embedd the elastic graph}
\usage{
CrossEmbedment(X_Source, X_Target, TargetPG, TrimmingRadius = Inf)
}
\arguments{
\item{X_Source}{}

\item{X_Target}{}

\item{TargetPG}{}

\item{TrimmingRadius}{}
}
\description{
Cross embedd the elastic graph
}
