% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{curve_data}
\alias{curve_data}
\title{Points on a tree}
\format{A data frame with 500 rows and 3 variables}
\description{
An example dataset containing 3d points distribued on a curve
}
\keyword{data}
