% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{DecodeElasticMatrix}
\alias{DecodeElasticMatrix}
\title{Converts ElasticMatrix into a set of edges and vectors of elasticities for edges and stars}
\usage{
DecodeElasticMatrix(ElasticMatrix)
}
\arguments{
\item{ElasticMatrix}{an e-by-e elastic matrix}
}
\value{
a list with three elements: a matrix with the edges (Edges), a vector of lambdas (Lambdas), and a vector of Mus (Mus)
}
\description{
Converts ElasticMatrix into a set of edges and vectors of elasticities for edges and stars
}
