% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVSEdges.R
\name{DistanceOnGraph}
\alias{DistanceOnGraph}
\title{Compute "on the graph" distance}
\usage{
DistanceOnGraph(X, TargetPG)
}
\arguments{
\item{X}{}

\item{TargetPG}{}
}
\description{
Compute "on the graph" distance
}
\examples{


library(ElPiGraph.R)
PG <- computeElasticPrincipalTree(X = tree_data, NumNodes = 50, InitNodes = 2, verbose = TRUE)
Dist <- DistanceOnGraph(X = tree_data, TargetPG = PG[[1]])
HC <- hclust(as.dist(Dist$FullDist))
GG <- cutree(HC, k = 10)

PlotPG(X = tree_data, TargetPG = PG[[1]], GroupsLab = factor(GG))

CompleteDist <- Dist$FullDist + Dist$DistFromGraph
HC <- hclust(as.dist(CompleteDist))
GGFull <- cutree(HC, k = 10)

PlotPG(X = tree_data, TargetPG = PG[[1]], GroupsLab = factor(GGFull))

table(GG, GGFull)


}
