% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVSEdges.R
\name{EdgeCatAssociation}
\alias{EdgeCatAssociation}
\title{Title}
\usage{
EdgeCatAssociation(X, TargetPG, GroupsLab = NULL)
}
\arguments{
\item{X}{}

\item{TargetPG}{}

\item{GroupsLab}{}
}
\description{
Title
}
