% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Encode2ElasticMatrix}
\alias{Encode2ElasticMatrix}
\title{Create an Elastic matrix from a set of edges}
\usage{
Encode2ElasticMatrix(Edges, Lambdas, Mus)
}
\arguments{
\item{Edges}{an e-by-2 matrix containing the index of the edges connecting the nodes}

\item{Lambdas}{the lambda parameters. Either a single value (which will be used for all the edges),
or a vector containing the values for each edge}

\item{Mus}{the mu parameters. Either a single value (which will be used for all the nodes),
or a vector containing the values for each node}
}
\value{
the elastic matrix
}
\description{
Create an Elastic matrix from a set of edges
}
