% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.structure.R
\name{ExploreBranching}
\alias{ExploreBranching}
\title{Title}
\usage{
ExploreBranching(X, TargetPG, GroupsLab = NULL, Partition = NULL,
  MaxNodes = Inf, nReps = 100, Prob = 0.9, PInf = 1, MaxNodeDist = 2,
  LabMult = 0.5)
}
\arguments{
\item{X}{}

\item{TargetPG}{}

\item{GroupsLab}{}

\item{Partition}{}

\item{MaxNodes}{}

\item{nReps}{}

\item{Prob}{}

\item{PInf}{}

\item{MaxNodeDist}{}
}
\description{
Title
}
\examples{

}
