% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{ExpOnPath}
\alias{ExpOnPath}
\title{Title}
\usage{
ExpOnPath(ExpData, Group = NULL, Path, TargetPG, Partition, PrjStr,
  Main = "", genes = 4)
}
\arguments{
\item{ExpData}{}

\item{TargetPG}{}

\item{Partition}{}

\item{PrjStr}{}

\item{Main}{}

\item{genes}{}

\item{Paths}{}
}
\description{
Title
}
