% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologies.R
\name{ExtendLeaves}
\alias{ExtendLeaves}
\title{Extend leaves with additional nodes}
\usage{
ExtendLeaves(X, TargetPG, Mode = "QuantCentroid", ControlPar = 0.9,
  LeafIDs = NULL, TrimmingRadius = Inf, PlotSelected = TRUE)
}
\arguments{
\item{X}{numeric matrix, the data matrix}

\item{TargetPG}{list, the ElPiGraph structure to extend}

\item{Mode}{string, the mode used to extend the graph. "QuantCentroid" and "WeigthedCentroid" are currently implemented}

\item{ControlPar}{positive numeric, the paramter used to control the contribution of the different data points}

\item{LeafIDs}{integer vector, the id of nodes to extend. If NULL, all the vertices will be extended.}

\item{TrimmingRadius}{positive numeric, the trimming radius used to control distance}

\item{PlotSelected}{boolean, should a diagnostic plot be visualized}
}
\value{
The extended ElPiGraph structure

The value of ControlPar has a different interpretation depending on the valus of Mode. In each case, for only the extreme points,
i.e., the points associated with the leaf node that do not have a projection on any edge are considered.

If Mode = "QuantCentroid", for each leaf node, the extreme points are ordered by their distance from the node
and the centroid of the points farther away than the ControlPar is returned.

If Mode = "WeigthedCentroid", for each leaf node, a weight is computed for each points by raising the distance to the ControlPar power.
Hence, larger values of ControlPar result in a larger influence of points farther from the node
}
\description{
Extend leaves with additional nodes
}
\examples{

ExtStruct <- ExtendLeaves(X = tree_data, TargetPG = TreeEPG[[1]], Mode = "QuantCentroid", ControlPar = .5)
PlotPG(X = tree_data, TargetPG = ExtStruct)

ExtStruct <- ExtendLeaves(X = tree_data, TargetPG = TreeEPG[[1]], Mode = "QuantCentroid", ControlPar = .9)
PlotPG(X = tree_data, TargetPG = ExtStruct)

ExtStruct <- ExtendLeaves(X = tree_data, TargetPG = TreeEPG[[1]], Mode = "WeigthedCentroid", ControlPar = .2)
PlotPG(X = tree_data, TargetPG = ExtStruct)

ExtStruct <- ExtendLeaves(X = tree_data, TargetPG = TreeEPG[[1]], Mode = "WeigthedCentroid", ControlPar = .8)
PlotPG(X = tree_data, TargetPG = ExtStruct)

}
