% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{FindAssocited}
\alias{FindAssocited}
\title{Find the points associted with the final ElPiGraph structure}
\usage{
FindAssocited(X, BootPG, TrimmingRadius = NULL)
}
\arguments{
\item{X}{numeric matrix (rows are points)}

\item{BootPG}{a list of ElPiGraph structures}

\item{TrimmingRadius}{the trimming radius to use. If NULL (the default), the one provided by the single ElPiGraph structures will be used}
}
\description{
Find the points associted with the final ElPiGraph structure
}
