% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{getPrimitiveGraphStructureBarCode}
\alias{getPrimitiveGraphStructureBarCode}
\title{Title}
\usage{
getPrimitiveGraphStructureBarCode(ElasticMatrix)
}
\arguments{
\item{ElasticMatrix}{}
}
\description{
Title
}
