% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{GetProjectionUncertainty}
\alias{GetProjectionUncertainty}
\title{Compute projection-associted uncertainty}
\usage{
GetProjectionUncertainty(X, BootPG, Mode = "MedianDistPW", TargetPG = NULL)
}
\arguments{
\item{X}{numeric matrix containing the points to be projected}

\item{BootPG}{a list of ElPiGraph structures. The nodes dimensionality must
be compatible with the dimensioanlity of the data}

\item{Mode}{string, the mode used to compute points uncertainty. Currently
the following options are supported
\describe{
\item{"MedianDistPW"}{All the paiwise distances between the projections
of the same point across the different ElPiGraphs are calculated and the
median (per points) is returned}
\item{"MedianDistPW"}{All the paiwise distances between the projections
of the same point across the different ElPiGraphs are calculated and the
mean (per points) is returned}
\item{"MeanDistCentroid"}{The centroid of the projections of the same point
across the different ElPiGraphs are calculated and the mean (per point)
distance of the original points from the centroid is returned}
\item{"MeanDistCentroid"}{The centroid of the projections of the same point
across the different ElPiGraphs are calculated and the mean (per point)
distance of the original points from the centroid is returned}
}}

\item{TargetPG}{an optional parameter describing the target ElPiGraph. Currently unused.}
}
\value{
a numeric vector with the same length as the number of rows of X. Higher values indicates larger uncertainty.
}
\description{
Compute projection-associted uncertainty
}
\examples{

  nRep <- 50
  
  
  TreeEPG <- computeElasticPrincipalTree(X = tree_data, NumNodes = 70,
      drawAccuracyComplexity = FALSE, drawEnergy = FALSE,
      drawPCAView = FALSE,
      nReps = nRep, ProbPoint = .9,
      TrimmingRadius = Inf,
      ICOver = "DensityProb", DensityRadius = .2)
  
  PtUnc_1 <- GetProjectionUncertainty(X = tree_data, BootPG = TreeEPG[1:nRep], Mode = "MedianDistPW")
  PtUnc_2 <- GetProjectionUncertainty(X = tree_data, BootPG = TreeEPG[1:nRep], Mode = "MedianDistCentroid")
  PtUnc_3 <- GetProjectionUncertainty(X = tree_data, BootPG = TreeEPG[1:nRep], Mode = "MeanDistPW")
  PtUnc_4 <- GetProjectionUncertainty(X = tree_data, BootPG = TreeEPG[1:nRep], Mode = "MeanDistCentroid")
  
  pairs(cbind(PtUnc_1, PtUnc_2, PtUnc_3, PtUnc_4))
  
  PlotPG(X = tree_data, TargetPG = TreeEPG[[nRep+1]], BootPG = TreeEPG[1:nRep], GroupsLab = PtUnc_1)
  PlotPG(X = tree_data, TargetPG = TreeEPG[[nRep+1]],  GroupsLab = PtUnc_1, p.alpha = .9)

}
