% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{GrowPath}
\alias{GrowPath}
\title{Grow a branching path using walkers}
\usage{
GrowPath(Walkers, StepSD = 0.1, nSteps = 100, BranchProb = 0.01,
  MinAgeBr = 50, BrDim = 5)
}
\arguments{
\item{Walkers}{list, a list of walker as returned from the InizializeWalkers function}

\item{nSteps}{integer, the number of steps}

\item{BranchProb}{numeric between 0 and 1, the probability per walker of branching at each step}

\item{MinAgeBr}{integer, the minimal number of steps before a newly introduced walker will start branhing}

\item{BrDim}{integer, the number of dimensions affected during branching}

\item{StepSize}{positive numeric, the standard deviation associated with the movement
in each direction for each step}
}
\description{
Grow a branching path using walkers
}
\examples{

Walkers <- InizializeWalkers(nDim = 1000)

Data <- GrowPath(Walkers = Walkers, StepSize = 50, nSteps = 2000, BranchProb = .0015, MinAgeBr = 75, BrDim = 15)

}
