% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferTrimRadius.R
\name{InferTrimRadius}
\alias{InferTrimRadius}
\title{Title}
\usage{
InferTrimRadius(X, nPoints, nInt = 100, plotCurves = FALSE)
}
\arguments{
\item{nPoints}{}

\item{plotCurves}{}
}
\description{
Title
}
