% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{InizializeWalkers}
\alias{InizializeWalkers}
\title{Initialize a group of 'walkers'}
\usage{
InizializeWalkers(Number = 1, Dimensions = 3, MeanShift = 1,
  SdShift = 0.3)
}
\arguments{
\item{Number}{integer, the number of walkers to be inizialised}

\item{Dimensions}{integer, the number of dimensions of the walkers}

\item{MeanShift}{numeric, the mean shift in each dimension for the walker}

\item{SdShift}{positive numeric, the standard deviation in the shift in each dimension for the walker}
}
\value{
A list with the walkers
}
\description{
Initialize a group of 'walkers'
}
\examples{

Walkers <- InizializeWalkers(Dimensions = 1000)

}
