% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.structure.R
\name{NodeDimensionality}
\alias{NodeDimensionality}
\title{Title}
\usage{
NodeDimensionality(X, TargetPG, Partition = NULL, MaxNodeDist = 4,
  nReps = 100, Prob = 0.9)
}
\arguments{
\item{X}{}

\item{TargetPG}{}

\item{Partition}{}

\item{MaxNodeDist}{}

\item{nReps}{}

\item{Prob}{}
}
\description{
Title
}
