% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{PartitionData}
\alias{PartitionData}
\title{Assign data points to a set of nodes}
\usage{
PartitionData(X, NodePositions, SquaredX, TrimmingRadius = Inf, nCores = 1)
}
\arguments{
\item{X}{an n-by-m numeric matrix with the coordinates of the data points}

\item{NodePositions}{an k-by-m numeric matrix with the coordiante of the nodes}

\item{SquaredX}{the sum by row of the squared positions rowSums(X^2)}

\item{TrimmingRadius}{Maximum distance of data points to the nodes to assign a point to a node}

\item{nCores}{number of cores to use.}
}
\value{
A list containing two components: Partition (containing the associated nodes) and 
Dists (containing the squqred distance from the node)
}
\description{
Assign data points to a set of nodes
}
