% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMSDEnergyPlot}
\alias{plotMSDEnergyPlot}
\title{Plot the MSD VS Energy plot}
\usage{
plotMSDEnergyPlot(ReportTable, Main = "")
}
\arguments{
\item{Main}{string, title of the plot}

\item{PrintGraph}{a struct returned by computeElasticPrincipalGraph}
}
\value{
a ggplot plot
}
\description{
Plot the MSD VS Energy plot
}
