% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotPG}
\alias{PlotPG}
\title{Plot data and principal graph(s)}
\usage{
PlotPG(X, TargetPG, BootPG = NULL, PGCol = "EPG",
  PlotProjections = "none", GroupsLab = NULL, PointViz = "points",
  Main = "", p.alpha = 0.3, PointSize = NULL, NodeLabels = NULL,
  LabMult = 1, Do_PCA = TRUE, DimToPlot = c(1, 2))
}
\arguments{
\item{X}{numerical 2D matrix, the n-by-m matrix with the position of n m-dimensional points}

\item{TargetPG}{the main principal graph to plot}

\item{BootPG}{A list of principal graphs that will be considered as bostrapped curves}

\item{PGCol}{string, the label to be used for the main principal graph}

\item{PlotProjections}{string, the plotting mode for the node projection on the principal graph.
It can be "none" (no projections will be plotted), "onNodes" (the projections will indicate how points are associated to nodes),
and "onEdges" (the projections will indicate how points are projected on edges or nodes of the graph)}

\item{GroupsLab}{factor or numeric vector. A vector indicating either a category or a numeric value associted with
each data point}

\item{PointViz}{}

\item{Main}{string, the title of the plot}

\item{p.alpha}{numeric between 0 and 1, the alpha value of the points. Lower values will prodeuce more transparet points}

\item{PointSize}{numeric vector, a vector indicating the size to be associted with each node of the graph}

\item{NodeLabels}{string vector, a vector indicating the label to be associted with each node of the graph}

\item{LabMult}{numeric, a multiplier controlling the size of node labels}

\item{Do_PCA}{bolean, should the node of the principal graph be used to derive principal component projections and
rotate the space? If TRUE the plots will use the "EpG PC" as dimensions, if FALSE, the original dimensions will be used.}

\item{DimToPlot}{a integer vector specifing the PCs (if Do_PCA=TRUE) or dimension (if Do_PCA=FALSE) to plot. All the
combination will be considered, so, for example, if DimToPlot = 1:3, three plot will be produced.}
}
\description{
Plot data and principal graph(s)
}
