% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPieNet}
\alias{plotPieNet}
\title{Plot a graph with pie chart associated with each node}
\usage{
plotPieNet(X, TargetPG, GroupsLab = NULL, Partition = NULL,
  TrimmingRadius = Inf, Graph = NULL, LayOut = "nicely", Main = "",
  ScaleFunction = sqrt, NodeSizeMult = 1, ColCat = NULL, PlotNet = TRUE,
  DirectionMat = NULL, Thr = 0.05, Arrow.size = 0.5, Leg.cex = 1)
}
\arguments{
\item{X}{numerical 2D matrix, the n-by-m matrix with the position of n m-dimensional points}

\item{TargetPG}{the main principal graph to plot}

\item{GroupsLab}{string factor, a vector indicating the category of each data point}

\item{Partition}{}

\item{TrimmingRadius}{}

\item{PlotNet}{}

\item{Leg.cex}{}
}
\description{
Plot a graph with pie chart associated with each node
}
