% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{PlotShiny}
\alias{PlotShiny}
\title{Auxiliary function used by the shiny interface}
\usage{
PlotShiny(BaseData, TargetPG, CombPCA, Selected, AssociatedPoints = FALSE,
  PGCol = "EpG", GroupsLab = NULL, PointSize = 1, p.alpha = 0.4,
  Main = "")
}
\arguments{
\item{BaseData}{}

\item{TargetPG}{}

\item{CombPCA}{}

\item{Selected}{}

\item{PGCol}{}

\item{GroupsLab}{}

\item{PointSize}{}

\item{p.alpha}{}

\item{Main}{}
}
\description{
Auxiliary function used by the shiny interface
}
