% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{PrimitiveElasticGraphEmbedment}
\alias{PrimitiveElasticGraphEmbedment}
\title{Function fitting a primitive elastic graph to the data}
\usage{
PrimitiveElasticGraphEmbedment(X, NodePositions, ElasticMatrix,
  MaxNumberOfIterations, TrimmingRadius, eps, Mode = 1,
  MinimizingEnergy = "Base", FinalEnergy = "Base", SquaredX = NULL,
  verbose = FALSE, FastSolve = FALSE, DisplayWarnings = FALSE,
  alpha = 1, beta = 1, prob = 1)
}
\arguments{
\item{X}{is n-by-m matrix containing the positions of the n points in the m-dimensional space}

\item{NodePositions}{is k-by-m matrix of positions of the graph nodes in the same space as X}

\item{ElasticMatrix}{is a k-by-k symmetric matrix describing the connectivity and the elastic
properties of the graph. Star elasticities (mu coefficients) are along the main diagonal
(non-zero entries only for star centers), and the edge elasticity moduli are at non-diagonal elements.}

\item{MaxNumberOfIterations}{is an integer number indicating the maximum number of iterations for the EM algorithm}

\item{TrimmingRadius}{is a real value indicating the trimming radius, a parameter required for robust principal graphs
(see https://github.com/auranic/Elastic-principal-graphs/wiki/Robust-principal-graphs)}

\item{eps}{a real number indicating the minimal relative change in the nodenpositions
to be considered the graph embedded (convergence criteria)}

\item{Mode}{integer, the energy mode. It can be 1 (difference is computed using the position of the nodes) and
2 (difference is computed using the changes in elestic energy of the configuraztions)}

\item{MinimizingEnergy}{string indicating the elastic emergy type to minimize if Mode = 2. Currently it can be "Base" or "Penalized"}

\item{FinalEnergy}{string indicating the final elastic emergy associated with the configuration. Currently it can be "Base" or "Penalized"}

\item{SquaredX}{the sum (by node) of X squared. It not specified, it will be calculated by the fucntion}

\item{verbose}{is a boolean indicating whether diagnostig informations should be plotted}

\item{FastSolve}{boolean, shuold the Fastsolve of Armadillo by enabled?}

\item{DisplayWarnings}{boolean, should warning about convergence be displayed?}

\item{alpha}{positive numeric, the value of the alpha parameter of the penalized elastic energy}

\item{prob}{numeric between 0 and 1. If less than 1 point will be sampled at each iteration. Prob indicate the probability of
using each points. This is an *experimental* feature, which may helps speeding up the computation if a large number of points is present.}

\item{betapositive}{numeric, the value of the beta parameter of the penalized elastic energy}
}
\description{
Function fitting a primitive elastic graph to the data
}
