% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{RearrangePlot}
\alias{RearrangePlot}
\title{Interactivelly modify data}
\usage{
RearrangePlot(X, TargetPG, PGCol = "EpG", GroupsLab = NULL, PointSize = 1,
  p.alpha = 0.4, Main = "")
}
\arguments{
\item{X}{}

\item{TargetPG}{}

\item{PGCol}{}

\item{GroupsLab}{}

\item{PointSize}{}

\item{p.alpha}{}

\item{Main}{}
}
\description{
Interactivelly modify data
}
