% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{ReportOnPrimitiveGraphEmbedment}
\alias{ReportOnPrimitiveGraphEmbedment}
\title{Title}
\usage{
ReportOnPrimitiveGraphEmbedment(X, NodePositions, ElasticMatrix,
  PartData = NULL, ComputeMSEP = FALSE)
}
\arguments{
\item{X}{}

\item{NodePositions}{}

\item{ElasticMatrix}{}

\item{PartData}{}

\item{ComputeMSEP}{}
}
\description{
Title
}
