% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.selection.R
\name{SelectDimensions}
\alias{SelectDimensions}
\title{Title}
\usage{
SelectDimensions(X, Partition, Edges, ProjStruct, EdgeSeq, Mode = "CV",
  AggFun = min, Span = 0.75, nCores = 1)
}
\arguments{
\item{X}{}

\item{Partition}{}

\item{Edges}{}

\item{ProjStruct}{}

\item{EdgeSeq}{}

\item{Mode}{string, the mode to compute dimension-associated pseutodime adherence statistics. It can be
"CV", "SmoothOnPath", or "LinearOnPath"}

\item{AggFun}{}

\item{Span}{}

\item{nCores}{}
}
\description{
Title
}
