% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlterStructure.R
\name{ShiftBranching}
\alias{ShiftBranching}
\title{Move branching nodes to areas with higher point density}
\usage{
ShiftBranching(X, TargetPG, SelectionMode = "NodePoints",
  DensityRadius = NA, MaxShift = 3, Compensate = FALSE, BrIds = NULL,
  TrimmingRadius = Inf)
}
\arguments{
\item{X}{numeric matrix, the data matrix}

\item{TargetPG}{list, the ElPiGraph structure to extend}

\item{SelectionMode}{string, the mode to use to shift the branching points. The "NodePoints" and "NodeDensity" modes are currently supported}

\item{DensityRadius}{positive numeric, the radius to be used when computing point density if SelectionMode = "NodeDensity"}

\item{MaxShift}{positive integer, the maxium distance (as number of edges) to consider when exploring the branching point neighborhood}

\item{Compensate}{booelan, should new points be included to compensate for olter one being removed (currently not implemented)}

\item{BrIds}{integer vector, the id of the branching points to consider. Id not associted with node possessing degree > 2 will be ignored}

\item{TrimmingRadius}{positive numeric, the trimming radius used to control distance}
}
\value{
a list with two components: NodePositions (Containing the new nodes positions) and Edges (containing the new edges)

The function explore the neighborhood of branching point for nodes with higher point density. It such point is found, to graph will be
modified so that the new found node will be the new branching point of the neighborhood.
}
\description{
Move branching nodes to areas with higher point density
}
