% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.structure.R
\name{StudyBranching}
\alias{StudyBranching}
\title{Title}
\usage{
StudyBranching(X, TargetPG, GroupsLab = NULL, Partition = NULL,
  MaxNodes = Inf, nReps = 100, Prob = 0.9, PInf = 1, MaxNodeDist = 4,
  DiagPlot = FALSE)
}
\arguments{
\item{X}{}

\item{TargetPG}{}

\item{GroupsLab}{}

\item{Partition}{}

\item{MaxNodes}{}

\item{nReps}{}

\item{Prob}{}

\item{PInf}{}

\item{MaxNodeDist}{}

\item{DiagPlot}{}
}
\description{
Title
}
