% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverDispersion.R
\name{StudyOverDispersion}
\alias{StudyOverDispersion}
\title{Title}
\usage{
StudyOverDispersion(X, NumNodes, Subset, Samples = 100, Type = "Curve",
  Exclusive = FALSE, ...)
}
\arguments{
\item{X}{}

\item{NumNodes}{}

\item{Subset}{}

\item{Samples}{}

\item{Type}{}

\item{Exclusive}{}

\item{...}{}
}
\description{
Title
}
