% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tree_data-name}
\alias{tree_data-name}
\title{Points on a tree}
\format{A data frame with 492 rows and 3 variables}
\description{
An example dataset containing 3d points distribued on a branching tree
}
\keyword{data}
