% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMAdjustment.R
\name{AdjustByConstant}
\alias{AdjustByConstant}
\title{Adjust elastic matrix by dividing the values of lambda and mu of stars}
\usage{
AdjustByConstant(GraphInfo, k = 2, divMu = 10, divLambda = 10)
}
\arguments{
\item{GraphInfo}{the elpigraph structure updated}

\item{k}{the largest orger of strars to leave unadjusted. e.g., if k = 2 only branching points will be adjusted}

\item{divMu}{the value used to divide the mu coefficients}

\item{divLambda}{the value used to divide the lambda coefficients}
}
\description{
Adjust elastic matrix by dividing the values of lambda and mu of stars
}
