% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ComputeRelativeChangeOfNodePositions}
\alias{ComputeRelativeChangeOfNodePositions}
\title{Replaced by a C++ function}
\usage{
ComputeRelativeChangeOfNodePositions(NodePositions, NewNodePositions,
  Mode = 1, BranchingFee = 0)
}
\arguments{
\item{NodePositions}{a k-by-m numeric matrix with the coordiantes of the nodes in the old configuration}

\item{NewNodePositions}{a k-by-m numeric matrix with the coordiantes of the nodes in the new configuration}

\item{Mode}{an integer indicating the modality used to compute the difference (currently only 1 is an accepted value)}

\item{BranchingFee}{currenty not used}

\item{X}{an n-by-m numeric matrix with the coordinates of the data points}
}
