% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVSEdges.R
\name{EdgeCatAssociation}
\alias{EdgeCatAssociation}
\title{Title}
\usage{
EdgeCatAssociation(X, TargetPG, GroupsLab = NULL, TrimmingRadius = Inf)
}
\arguments{
\item{X}{a data matrix}

\item{TargetPG}{an ElPiGraph object}

\item{GroupsLab}{a vector of labels for the point of the matrix}
}
\value{
a list with assocition tables and Chi-squred test for category associtions with Edges+Nodes and Nodes
}
\description{
Title
}
