% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologies.R
\name{generateInitialConfiguration}
\alias{generateInitialConfiguration}
\title{Produce an initial graph with a given structure}
\usage{
generateInitialConfiguration(X, Nodes, Configuration = "Line",
  DensityRadius = NULL, MaxPoints = 25000, PCADensity = TRUE,
  CenterDataDensity = TRUE)
}
\arguments{
\item{X}{numerical 2D matrix, the n-by-m matrix with the position of n m-dimensional points}

\item{Nodes}{integer, number of nodes of the graph}

\item{Configuration}{string, type of graph to return. It can be "Line", "Circle", or "Density"}

\item{DensityRadius}{numeric, the radius used to estimate local density. This need to be set when Configuration is equal to "Density"}

\item{MaxPoints}{integer, the maximum number of points for which the local density will be estimated. If the number of data points is
larger than MaxPoints, a subset of the original points will be sampled}

\item{PCADensity}{boolean, should PCA be applied to the data before computing the most dense area}
}
\description{
Produce an initial graph with a given structure
}
