% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{GenertateConsensusGraph}
\alias{GenertateConsensusGraph}
\title{Obtain a consens tree from bootstrapped data}
\usage{
GenertateConsensusGraph(BootPG, Mode = "NodeDist", CollapseRadius = NULL,
  Clusters = 100, MinEdgMult = 2, MinNodeMult = 2,
  RemoveIsolatedNodes = TRUE, OnlyLCC = FALSE, PlotResult = TRUE)
}
\arguments{
\item{BootPG}{list of ElPiGraph structures containing the bootstrapped data}

\item{Mode}{string, the mode to use}

\item{MinEdgMult}{integer, minimal multiplicity of the edges of the consensus graph}

\item{MinNodeMult}{integer, minimal multiplicity of the nodes of the consensus graph}

\item{RemoveIsolatedNodes}{}

\item{OnlyLCC}{}

\item{PlotResult}{}

\item{MinTol}{numeric, the smallest distance for two edges to be collapsed}

\item{NodesInflation}{integer, the}
}
\description{
Obtain a consens tree from bootstrapped data
}
