% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{GenertateConsensusGraph}
\alias{GenertateConsensusGraph}
\title{Obtain a consens tree from bootstrapped data}
\usage{
GenertateConsensusGraph(BootPG, Mode = "NodeDist", MinTol = 0.3,
  MinEdgMult = 2, MinNodeMult = 2, RemoveIsolatedNodes = TRUE)
}
\arguments{
\item{BootPG}{list of ElPiGraph structures containing the bootstrapped data}

\item{Mode}{string, the mode to use}

\item{MinTol}{numeric, the smallest distance for two edges to be collapsed}

\item{MinEdgMult}{integer, minimal multiplicity of the edges of the consensus graph}

\item{MinNodeMult}{integer, minimal multiplicity of the nodes of the consensus graph}
}
\description{
Obtain a consens tree from bootstrapped data
}
