% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{getPseudotime}
\alias{getPseudotime}
\title{Obtain the pseudotime associted with a path}
\usage{
getPseudotime(ProjStruct, NodeSeq)
}
\arguments{
\item{EdgeSeq}{}
}
\description{
Obtain the pseudotime associted with a path
}
