% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferTrimRadius.R
\name{InferTrimRadius}
\alias{InferTrimRadius}
\title{Title}
\usage{
InferTrimRadius(X, nPoints = NULL, nInt = 100, plotCurves = FALSE,
  n.cores = 1, ClusType = "FORK")
}
\arguments{
\item{nInt}{}

\item{n.cores}{}

\item{ClusType}{}
}
\description{
Title
}
