% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.selection.R
\name{MeasureSmoothness}
\alias{MeasureSmoothness}
\title{Measure smoothness of the differnt feature of the data}
\usage{
MeasureSmoothness(X, Paths, TargetPG, SmoothMode = "lowess",
  CollMode = "var", Partition = NULL, PrjStr = NULL, Randomize = 0,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Measure smoothness of the differnt feature of the data
}
