% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{project_point_onto_graph}
\alias{project_point_onto_graph}
\title{Title}
\usage{
project_point_onto_graph(X, NodePositions, Edges, Partition = NULL)
}
\arguments{
\item{Partition}{}
}
\description{
Title
}
